Shuffling folder structure; changed from sql to nosql
