+++ package.json
@@ -21,9 +21,7 @@
     "jsonwebtoken": "^8.5.1",
     "lodash": "^4.17.15",
     "moment": "^2.24.0",
-    "pg": "^7.17.0",
-    "pg-hstore": "^2.3.3",
-    "sequelize": "^5.21.3"
+    "mongoose": "^5.9.1"
   },
   "devDependencies": {
     "cross-env": "^6.0.3",
